# Cache the Git remote URL to speed up rendering hyperlinks