--- a/src/features/hyperlinks.rs
+++ b/src/features/hyperlinks.rs
@@ -21,9 +21,7 @@ pub fn make_feature() -> Vec<(String, OptionValueFunction)> {
 
 #[cfg(test)]
 pub fn remote_from_config(_: &Option<&GitConfig>) -> Option<GitRemoteRepo> {
-    Some(GitRemoteRepo::GitHub {
-        slug: "dandavison/delta".to_string(),
-    })
+    GitRemoteRepo::for_testing()
 }
 
 #[cfg(not(test))]
@@ -41,6 +39,7 @@ lazy_static! {
 pub fn format_commit_line_with_osc8_commit_hyperlink<'a>(
     line: &'a str,
     config: &Config,
+    cached_repo: &Option<GitRemoteRepo>,
 ) -> Cow<'a, str> {
     // Given matches in a line, m = matches[0] and pos = 0: store line[pos..m.start()] first, then
     // store the T(line[m.start()..m.end()]) match transformation, then set pos = m.end().
@@ -80,7 +79,7 @@ pub fn format_commit_line_with_osc8_commit_hyperlink<'a>(
                     .with_input(line, &first_match, &mut matches);
             return Cow::from(result);
         }
-    } else if let Some(repo) = remote_from_config(&config.git_config()) {
+    } else if let Some(repo) = cached_repo {
         let mut matches = COMMIT_HASH_REGEX.find_iter(line);
         if let Some(first_match) = matches.next() {
             let result = HyperlinkCommits(|commit_hash| repo.format_commit_url(commit_hash))
@@ -144,27 +143,28 @@ pub mod tests {
 
     #[test]
     fn test_formatted_hyperlinks() {
+        let remote = GitRemoteRepo::for_testing();
         let config = make_config_from_args(&["--hyperlinks-commit-link-format", "HERE:{commit}"]);
-
         let line = "001234abcdf";
-        let result = format_commit_line_with_osc8_commit_hyperlink(line, &config);
+
+        let result = format_commit_line_with_osc8_commit_hyperlink(line, &config, &remote);
         assert_eq!(
             result,
             "\u{1b}]8;;HERE:001234abcdf\u{1b}\\001234abcdf\u{1b}]8;;\u{1b}\\",
         );
 
         let line = "a2272718f0b398e48652ace17fca85c1962b3fc22"; // length: 41 > 40
-        let result = format_commit_line_with_osc8_commit_hyperlink(line, &config);
+        let result = format_commit_line_with_osc8_commit_hyperlink(line, &config, &remote);
         assert_eq!(result, "a2272718f0b398e48652ace17fca85c1962b3fc22",);
 
         let line = "a2272718f0+b398e48652ace17f,ca85c1962b3fc2";
-        let result = format_commit_line_with_osc8_commit_hyperlink(line, &config);
+        let result = format_commit_line_with_osc8_commit_hyperlink(line, &config, &remote);
         assert_eq!(result, "\u{1b}]8;;HERE:a2272718f0\u{1b}\\a2272718f0\u{1b}]8;;\u{1b}\\+\u{1b}]8;;\
         HERE:b398e48652ace17f\u{1b}\\b398e48652ace17f\u{1b}]8;;\u{1b}\\,\u{1b}]8;;HERE:ca85c1962b3fc2\
         \u{1b}\\ca85c1962b3fc2\u{1b}]8;;\u{1b}\\");
 
         let line = "This 01234abcdf Hash";
-        let result = format_commit_line_with_osc8_commit_hyperlink(line, &config);
+        let result = format_commit_line_with_osc8_commit_hyperlink(line, &config, &remote);
         assert_eq!(
             result,
             "This \u{1b}]8;;HERE:01234abcdf\u{1b}\\01234abcdf\u{1b}]8;;\u{1b}\\ Hash",
@@ -172,7 +172,7 @@ pub mod tests {
 
         let line =
             "Another 01234abcdf hash but also this one: dc623b084ad2dd14fe5d90189cacad5d49bfbfd3!";
-        let result = format_commit_line_with_osc8_commit_hyperlink(line, &config);
+        let result = format_commit_line_with_osc8_commit_hyperlink(line, &config, &remote);
         assert_eq!(
             result,
             "Another \u{1b}]8;;HERE:01234abcdf\u{1b}\\01234abcdf\u{1b}]8;;\u{1b}\\ hash but \
@@ -181,7 +181,7 @@ pub mod tests {
         );
 
         let line = "01234abcdf 03043baf30 12abcdef0 12345678";
-        let result = format_commit_line_with_osc8_commit_hyperlink(line, &config);
+        let result = format_commit_line_with_osc8_commit_hyperlink(line, &config, &remote);
         assert_eq!(
             result,
             "\u{1b}]8;;HERE:01234abcdf\u{1b}\\01234abcdf\u{1b}]8;;\u{1b}\\ \u{1b}]8;;\
@@ -194,9 +194,10 @@ pub mod tests {
     fn test_hyperlinks_to_repo() {
         let mut config = make_config_from_args(&["--hyperlinks"]);
         config.git_config = GitConfig::for_testing();
+        let remote = GitRemoteRepo::for_testing();
 
         let line = "This a589ff9debaefdd delta commit";
-        let result = format_commit_line_with_osc8_commit_hyperlink(line, &config);
+        let result = format_commit_line_with_osc8_commit_hyperlink(line, &config, &remote);
         assert_eq!(
             result,
             "This \u{1b}]8;;https://github.com/dandavison/delta/commit/a589ff9debaefdd\u{1b}\
@@ -205,7 +206,7 @@ pub mod tests {
 
         let line =
             "Another a589ff9debaefdd hash but also this one: c5696757c0827349a87daa95415656!";
-        let result = format_commit_line_with_osc8_commit_hyperlink(line, &config);
+        let result = format_commit_line_with_osc8_commit_hyperlink(line, &config, &remote);
         assert_eq!(
             result,
             "Another \u{1b}]8;;https://github.com/dandavison/delta/commit/a589ff9debaefdd\