--- a/src/handlers/commit_meta.rs
+++ b/src/handlers/commit_meta.rs
@@ -37,10 +37,12 @@ impl StateMachine<'_> {
                 features::hyperlinks::format_commit_line_with_osc8_commit_hyperlink(
                     &self.line,
                     self.config,
+                    &self.git_remote_repo,
                 ),
                 features::hyperlinks::format_commit_line_with_osc8_commit_hyperlink(
                     &self.raw_line,
                     self.config,
+                    &self.git_remote_repo,
                 ),
             )
         } else {