This commit speeds up the rendering of hyperlinks by ~47x by caching the Git repos remote URL instead of fetching it each time a hyperlink is rendered.

Fixes #1939

### Note

I am very new to Rust so I am very open to feedback and completely understand if you want to rewrite or abandon this entire PR (especially if for some reason the remote URL should not be cached). Additionally, I think the [`GitConfig`](https://github.com/charlievieth/delta/blob/f4f71c15a110c77853c9afc8d77a6600394c55a5/src/git_config/mod.rs#L12-L20) should be responsible for storing the cached remote, but struggled to make that work due to borrowing, lifetimes, mutability, etc.

### Benchmarks

The below benchmarks were ran in the root of [Linux v6.12](https://github.com/torvalds/linux/tree/v6.12). We use `head -c $((128 * 1024 * 1024))` to limit the amount of Git log data since otherwise benchmarking `delta` before this change takes a very long time.

**TLDR:** This change makes processing 128Mib of Git log output take 3.7 seconds instead of 176.1 seconds.

#### Without git remote caching

```bash
$ DELTA_PAGER='bash -euo pipefail -c "time head -c $((128 * 1024 * 1024)) | pv --stats --output=/dev/null"' git -c 'pager.log=delta --hyperlinks' log
 128MiB 0:02:56 [ 744KiB/s]
rate min/avg/max/mdev = 654979.457/764202.554/1108112.678/35667.893 B/s

real	2m56.063s
user	0m0.736s
sys	0m5.530s
```

#### With git remote caching (this change)

```sh
$ PATH="{PATH_TO_NEW_DELTA}:${PATH}"
$ DELTA_PAGER='bash -c "time head -c $((128 * 1024 * 1024)) | pv --stats --output=/dev/null"' \
    git -c 'pager.log=delta --hyperlinks' log
 128MiB 0:00:03 [34.6MiB/s]
rate min/avg/max/mdev = 35979500.033/36287504.626/36646720.334/263641.477 B/s

real	0m3.700s
user	0m0.205s
sys	0m1.690s
```