@th1000s 2025-01-12
Thanks for reporting this!

Caching is fine, the slowdown is from libgit2 being overly correct and assuming that the remote url could change while the program is running.
The PR can indeed be made simpler by using a `RefCell` to add interior mutability to the `GitConfig`, i.e. `cached_remote: RefCell<Option<Option<String>>>` with an not-yet-cached option layer, and another for the result of `.find_remote(..).ok()` - or maybe a custum enum is clearer.