@charlievieth 2025-08-14
@th1000s Took a while until I had some spare cycles to look at this again, but following your guidance I was able to use a `OnceCell` to embed the `GitRemoteRepo` struct into the `GitConfig` struct. Thank you again for your suggestions and would love another review if you have the time.