@charlievieth 2025-09-05
@dandavison 1) Thank you for delta - it's incredible 2) Any chance you could enable the tests?