@charlievieth 2025-09-05
@dandavison Thank you and I completely understand (I've only recently had the cycles to address my own OSS commits and projects).

The `clippy` [failure](https://github.com/dandavison/delta/actions/runs/16953832997/job/49654943580?pr=1940) that I get on this branch/PR are identical to the one I get when running `make lint` (aka `cargo clippy`) against the main branch (d5e0565cbfa47acde98d41d8777ace5d1bc4d690). I'm fairly new to Rust so happy to look into the failure myself and include it into this branch or create a preceding PR (to merge before this one) that addresses the issue.

For context my rust toolchain is:
* rustc `1.89.0` (29483883e 2025-08-04) (Homebrew)
* clippy `0.1.89`

## Script used to check the delta of `cargo clippy`

```bash
#!/usr/bin/env bash

set -euo pipefail

TMP=$(mktemp -d)
TMP_MAIN="${TMP}/clippy.main.txt"
TMP_HYPER="${TMP}/clippy.hyperlink.txt"

if ! git remote -v | grep -q '^charlievieth'; then
    git remote add charlievieth https://github.com/charlievieth/delta.git
fi
git fetch --all

git checkout main
cargo clippy  &> "${TMP_MAIN}" || true

git checkout cev/hyperlink
cargo clippy  &> "${TMP_HYPER}" || true

(
    cd "${TMP}"
    git diff --no-index "${TMP_MAIN}" "${TMP_HYPER}"
)
````

### Output (nothing applicable AFAIK):
```diff
diff --git a/var/folders/lz/4n_p3kf54zn1r8y3f_hvs94c0000gn/T/tmp.yJewF8SYWG/clippy.main.txt b/var/folders/lz/4n_p3kf54zn1r8y3f_hvs94c0000gn/T/tmp.yJewF8SYWG/clippy.hyperlink.txt
index 65e5421..4d0cc16 100644
--- a/var/folders/lz/4n_p3kf54zn1r8y3f_hvs94c0000gn/T/tmp.yJewF8SYWG/clippy.main.txt
+++ b/var/folders/lz/4n_p3kf54zn1r8y3f_hvs94c0000gn/T/tmp.yJewF8SYWG/clippy.hyperlink.txt
@@ -71,4 +71,4 @@ help: use `'_` for type paths
     |                                                                              ++++

 warning: `git-delta` (bin "delta") generated 5 warnings
-    Finished `dev` profile [unoptimized + debuginfo] target(s) in 0.68s
+    Finished `dev` profile [unoptimized + debuginfo] target(s) in 0.69s
```