@charlievieth 2025-09-09
> If you'd like to deal with the clippy complaints then that would be great! I think the best way would be to open a PR for the clippy-related fixes against `main` and then rebase this on that.

Created https://github.com/dandavison/delta/pull/2038 to address the clippy complaints.